% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/entropyFunction.R
\name{ber_analyze_dir_affect}
\alias{ber_analyze_dir_affect}
\title{Estimate Behavioral Entropy Rate based upon Video Data Files for a Directory}
\usage{
ber_analyze_dir_affect(
  dir_loc,
  tactile_padding = 1,
  auditory_padding = 1,
  behavior_types = list(mom_auditory_types = c("Vocal"), mom_tactile_types =
    c("TouchBaby", "HoldingBaby"), mom_visual_types = c("ManipulatingObject"),
    baby_visual_types = c("LookAtMomActivity"), missing_types = c("CantTellHolding",
    "ActivityNotVisible", "CantTellLooking", "CantTellAffect"), positive = c("positive"),
    negative = c("negative"), neutral = c("neutral")),
  missing_threshold = 0.1,
  log_file = paste(Sys.Date(), "-ber-logfile.txt", sep = "")
)
}
\arguments{
\item{dir_loc}{directory of files}

\item{tactile_padding}{right padding adjustment to tactile events}

\item{auditory_padding}{right padding adjustment to auditory events}

\item{behavior_types}{dictionary of behavior types.  Required sections, mom_auditory_types, mom_tactile_types, mom_visual_types, baby_visual_types, pos, neg, neutral, missing_types}

\item{missing_threshold}{proportion of acceptable missing time}
}
\value{
Entropy rate estimates of an individual
}
\description{
Estimate Behavioral Entropy Rate based upon Video Data Files for a Directory
}
