% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/entropyFunction.R
\name{ber_analyze_dir_affect_and_autonomy}
\alias{ber_analyze_dir_affect_and_autonomy}
\title{Estimate Behavioral Entropy Rate based upon Video Data Files for a Directory About Sensory Signals Affect and Autonomy}
\usage{
ber_analyze_dir_affect_and_autonomy(
  dir_loc,
  tactile_padding = 1,
  auditory_padding = 1,
  behavior_types = list(missing_types = c("CantTellAffect", "CantTellBehavior"), positive
    = c("positive"), negative = c("negative"), neutral = c("neutral"), autonomy_support =
    c("AutonomySupport"), intrusiveness = c("Intrusiveness"), neither = c("Neither")),
  missing_threshold = 0.1,
  log_file = paste(Sys.Date(), "-ber-logfile.txt", sep = "")
)
}
\arguments{
\item{dir_loc}{directory of files}

\item{tactile_padding}{right padding adjustment to tactile events}

\item{auditory_padding}{right padding adjustment to auditory events}

\item{behavior_types}{dictionary of behavior types.  Required sections, positive, negative, neutral,
autonomy support, intrusiveness, neither, missing_types}

\item{missing_threshold}{proportion of acceptable missing time}

\item{log_file}{file path for the log file to write out messages to}
}
\value{
Entropy rate estimates of an individual
}
\description{
Estimate Behavioral Entropy Rate based upon Video Data Files for a Directory About Sensory Signals Affect and Autonomy
}
