% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conversionFunction.R
\name{Boris2NoldusFileReShape}
\alias{Boris2NoldusFileReShape}
\title{Function to convert a single BORIS file into an ecber compatible excel file}
\usage{
Boris2NoldusFileReShape(rel)
}
\arguments{
\item{rel}{The individual CSV file that has been read in with read.csv()}
}
\value{
The final formatted data frame is returned
}
\description{
This function takes in a CSV file that has been read in with the read.csv() function, and returns the data frame of the ecber compatible output. The ConvertCSVtoExcel function uses this function as an intermediate step before writing out the data frame to an excel.
}
