% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reliabilityFunctions.R
\name{buildMatrix}
\alias{buildMatrix}
\title{Creating Agreement/Confusion Matrix between 2 coders with a tolerance of 0}
\usage{
buildMatrix(filepath1, filepath2, eventlist, file_seconds = 600)
}
\arguments{
\item{filepath1}{Filepath to first excel file.}

\item{filepath2}{Filepath to second excel file.}

\item{eventlist}{List of all states for each sensory signal, each list entry being a vector of the states per sensory signal.}

\item{file_seconds}{The total number of seconds coded for in each file.}
}
\value{
We return a list of two entries: the first being the confusion matrix and the second being the percent agreement metric.
}
\description{
This function takes in the 2 filepaths and the events that we are looking at for coder similarity. We create an agreement/confusion matrix comparing matches between codes in both files.
}
\examples{
buildMatrix(filepath1 = './6008BECEU.xlsx', filepath2 = './6008BECNR.xlsx', eventlist = list(c("positive","neutral","negative","CantTellAffect"), c("AutonomySupport","Neither","Intrusiveness","CantTellBehavior")), 300)
}
