% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kappaCalc.R
\name{calcCohensKappa}
\alias{calcCohensKappa}
\title{Calculating Cohen's Kappa, percent agreement, and percent by chance from a confusion matrix}
\usage{
calcCohensKappa(mat, total = 0)
}
\arguments{
\item{mat}{The first input is the confusion matrix without row or column sum values}

\item{total}{The sum of all values in the confusion matrix}
}
\value{
This function returns a data frame of all calculations.
}
\description{
This function returns Cohen's kappa, percent agreement, and percent by chance.
}
\examples{
calcCohensKappa(confusion_matrix, total)
}
