% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reliabilityFunctions.R
\name{calcKappaValuesFromMatrix}
\alias{calcKappaValuesFromMatrix}
\title{Calculating Kappa metrics and Percent Agreement From a Confusion Matrix}
\usage{
calcKappaValuesFromMatrix(matrixVals)
}
\arguments{
\item{matrixVals}{Confusion matrix without row and column sums}
}
\value{
We return a data frame of all the metrics we are calculating.
}
\description{
This function takes in as input a confusion matrix without the row or column sums, and calculates Cohen's kappa, unweighted and weighted kappa standard error, weighted kappa, percent agreement, and percent by chance values
}
\examples{
calcKappaValuesFromMatrix(agreement_matrix)
}
