% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/entropyFunction.R
\name{combined_entropy_output}
\alias{combined_entropy_output}
\title{Combined entropy output}
\usage{
combined_entropy_output(
  file_name,
  tactile_padding = 1,
  auditory_padding = 1,
  TAV_behavior_types = list(mom_auditory_types = c("Vocal"), mom_tactile_types =
    c("TouchBaby", "HoldingBaby"), mom_visual_types = c("ManipulatingObject"),
    baby_visual_types = c("LookAtMomActivity"), missing_types = c("CantTellHolding",
    "ActivityNotVisible", "CantTellLooking")),
  affect_behavior_types = list(mom_auditory_types = c("Vocal"), mom_tactile_types =
    c("TouchBaby", "HoldingBaby"), mom_visual_types = c("ManipulatingObject"),
    baby_visual_types = c("LookAtMomActivity"), positive = c("positive"), negative =
    c("negative"), neutral = c("neutral"), missing_types = c("CantTellHolding",
    "ActivityNotVisible", "CantTellLooking", "CantTellAffect")),
  autonomy_behavior_types = list(positive = c("positive"), negative = c("negative"),
    neutral = c("neutral"), autonomy_support = c("AutonomySupport"), intrusiveness =
    c("Intrusiveness"), neither = c("Neither"), missing_types = c("CantTellAffect",
    "CantTellBehavior")),
  missing_threshold = 0.1,
  order = 1
)
}
\arguments{
\item{file_name}{excel file to read in}

\item{tactile_padding}{default tactile padding is set to 1.0}

\item{auditory_padding}{default auditory padding is set to 1.0}

\item{TAV_behavior_types}{list of all the behavior types: Tactile Auditory Visual (TAV)}

\item{affect_behavior_types}{list of affect behavior types including TAV and positive, negative, neutral}

\item{autonomy_behavior_types}{list of autonomy behavior types including autonomy support, intrusiveness, and neither}

\item{missing_threshold}{default missing threshold value}
}
\value{
data frame of all of the entropy outputs for different subsets of the data
}
\description{
Combined entropy output
}
