% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/entropyFunction.R
\name{combined_entropy_output_dir}
\alias{combined_entropy_output_dir}
\title{This function goes through an entire directory and finds the combined entropy output for each file}
\usage{
combined_entropy_output_dir(
  dir_loc,
  log_file = paste(Sys.Date(), "-ber-logfile.txt", sep = ""),
  missing_threshold = 0.1,
  order = 1
)
}
\description{
This function goes through an entire directory and finds the combined entropy output for each file
}
