% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/entropyFunction.R
\name{combined_entropy_output_more_detail_dir}
\alias{combined_entropy_output_more_detail_dir}
\title{Combined entropy output for a directory with additional information}
\usage{
combined_entropy_output_more_detail_dir(
  dir_loc,
  log_file = paste(Sys.Date(), "-ber-logfile.txt", sep = ""),
  missing_threshold = 0.1,
  order = 1
)
}
\arguments{
\item{dir_loc}{directory path containing excel files to read in}

\item{log_file}{name of log file to write output messages to}

\item{missing_threshold}{default missing threshold value}

\item{order}{Markov model order value}
}
\value{
data frame of all entropy outputs for each file in the directory
}
\description{
This function is an extension of the combined_entropy_output_more_detail function except that it applies to an entire directory and creates a data frame with all the directory file information
}
