% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/entropyFunction.R
\name{combined_markov_lags}
\alias{combined_markov_lags}
\title{Combined entropy output for different markov model orders}
\usage{
combined_markov_lags(file_name, missing_threshold = 0.1, order_vals = 1)
}
\arguments{
\item{file_name}{excel file to read in}

\item{missing_threshold}{default missing threshold value}

\item{order_vals}{This can be a vector of different order values to include in the output, or a single order value}
}
\value{
data frame of all of the combined entropy outputs for different markov order values
}
\description{
This function takes in a list of orders for the transition matrix so we can compare detailed entropy rate information with different lag values. The entropy rates included are for all combinations of sensory signals including: TAV (Tactile, Auditory, and Visual), TAV and Affect, and Autonomy and Affect. Information about the number of transitions is also included.
}
