% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conversionFunction.R
\name{ConvertCSVtoEXCEL}
\alias{ConvertCSVtoEXCEL}
\title{Conversion from BORIS CSVs to ecber compatible excel files}
\usage{
ConvertCSVtoEXCEL(csv_dir)
}
\arguments{
\item{csv_dir}{The path to the directory containing the CSV files to convert.}
}
\value{
Nothing is returned but a new folder is created with the final excel files.
}
\description{
This function reads in a csv directory and creates a new sub-folder to contain the converted excel files
}
\examples{
ConvertCSVtoExcel('./path_to_csv_dir') creates a ConvertedToExcels sub-folder
}
