% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kappaCalc.R
\name{createPracticeExcel}
\alias{createPracticeExcel}
\title{This function creates a practice confusion matrix to calculate kappa values with.}
\usage{
createPracticeExcel(dimr = 5, dimc = 5, wd = getwd())
}
\arguments{
\item{dimr}{This value represents the fixed number of rows for the confusion matrix.}

\item{dimc}{This value represents the fixed number of columns for the confusion matrix.}

\item{wd}{The directory in which to write the practice excel file into}
}
\value{
This function doesn't return anything, but saves out the practice confusion matrix to the current directory.
}
\description{
This function creates a practice confusion matrix to calculate kappa values with.
}
\examples{
createPracticeExcel()
}
