% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reliabilityFunctions.R
\name{createReliabilityExcel}
\alias{createReliabilityExcel}
\title{Creating A Reliability Matrix Output File}
\usage{
createReliabilityExcel(
  save_dir,
  filename1,
  filename2,
  eventlist,
  file_seconds = 300
)
}
\arguments{
\item{save_dir}{Directory path to save the excel file into}

\item{eventlist}{List of all states for each sensory signal, each list entry being a vector of the states per sensory signal.}

\item{file_seconds}{The total number of seconds coded for in each file.}

\item{filepath1}{Filepath to first excel file.}

\item{filepath2}{Filepath to second excel file.}
}
\value{
Nothing is explicitly returned from this function, but the excel file is written and saved out into the output directory. The excel file contains all percent agreement and percent by chance values as well as kappa values.
}
\description{
This function takes in the 2 filepaths and a directory for the output excel file to be saved into. The reliability values calculated are all for fixed tolerance values of 0 and 1 for a direct comparison.
}
\examples{
createReliabilityExcel(save_dir, filepath1 = './6008BECEU.xlsx', filepath2 = './6008BECNR.xlsx', eventlist = list(c("positive","neutral","negative","CantTellAffect"), c("AutonomySupport","Neither","Intrusiveness","CantTellBehavior")), file_seconds = 300)
}
