% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reliabilityFunctions.R
\name{createReliabilityExcel2}
\alias{createReliabilityExcel2}
\title{Creating A Reliability Matrix Output File}
\usage{
createReliabilityExcel2(
  save_dir,
  save_as_fname,
  filename1,
  filename2,
  eventlist,
  tolerance = c(0, 1),
  file_seconds = 300
)
}
\arguments{
\item{save_dir}{Directory path to save the excel file into}

\item{eventlist}{List of all states for each sensory signal, each list entry being a vector of the states per sensory signal.}

\item{tolerance}{List of tolerance values to use for calculation}

\item{file_seconds}{The total number of seconds coded for in each file.}

\item{filepath1}{Filepath to first excel file.}

\item{filepath2}{Filepath to second excel file.}
}
\value{
Nothing is explicitly returned from this function, but the excel file is written and saved out into the output directory. The excel file contains all percent agreement and percent by chance values as well as kappa values for each of the tolerance values.
}
\description{
This function is the same as the createReliabilityExcel function except that we can now customize the tolerance values with a list of multiple values we initialize as the user. The tolerance values default to 0 and 1 if we don't change the input parameter.
}
\examples{
createReliabilityExcel(save_dir, filepath1 = './6008BECEU.xlsx', filepath2 = './6008BECNR.xlsx', eventlist = list(c("positive","neutral","negative","CantTellAffect"), c("AutonomySupport","Neither","Intrusiveness","CantTellBehavior")), file_seconds = 300)
}
