% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reliabilityFunctions.R
\name{createToleranceAndKappaValsFromFiles}
\alias{createToleranceAndKappaValsFromFiles}
\title{Calculating percent agreement and kappa values between 2 files}
\usage{
createToleranceAndKappaValsFromFiles(
  filepath1,
  filepath2,
  eventlist,
  tolerance,
  file_seconds = 300
)
}
\arguments{
\item{filepath1}{Filepath to first excel file.}

\item{filepath2}{Filepath to second excel file.}

\item{eventlist}{List of all states for each sensory signal, each list entry being a vector of the states per sensory signal.}

\item{tolerance}{Tolerance value for the amount of seconds that we buffer for comparing coder similarity, a tolerance of 0 requiring the same exact coding for each second}

\item{file_seconds}{The total number of seconds coded for in each file.}
}
\value{
This function returns a data frame with percent agreement, percent by chance, and various kappa values.
}
\description{
This function takes in the 2 filepaths, an event list, a tolerance value, and the total number of file seconds. An intermediate step of the function is the calculation of the confusion matrix, and this is used to calculate the percent agreement values along with the kappa values for comparing coder report similarity.
}
\examples{
createToleranceAndKappaValsFromFiles(filepath1, filepath2, eventlist, tolerance, file_seconds)
}
