% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reliabilityFunctions.R
\name{createToleranceAndKappaValsFromMatrix}
\alias{createToleranceAndKappaValsFromMatrix}
\title{Calculating percent agreement and kappa values from a confusion matrix}
\usage{
createToleranceAndKappaValsFromMatrix(output)
}
\arguments{
\item{output}{The confusion matrix already calculated between two files}
}
\value{
This function returns a data frame with percent agreement, percent by chance, and various kappa values.
}
\description{
This function takes in the confusion matrix already calculated from 2 excel files, and returns the kappa values, percent by chance, and percent agreement values.
}
\examples{
createToleranceAndKappaValsFromMatrix(confusion_matrix)
}
