% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kappaCalc.R
\name{individualCodeKappas}
\alias{individualCodeKappas}
\title{Calculating Individual Kappa Values for the K events which were coded for}
\usage{
individualCodeKappas(mat, total)
}
\arguments{
\item{mat}{The confusion matrix from both coders with the row and sum totals removed.}

\item{total}{The sum of all values in the matrix.}
}
\value{
A data frame with the individual code kappas for each event.
}
\description{
This function takes in a confusion matrix as well as the total of all values in the matrix and calculates the kappa for each of the individual events.
}
